# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "operationId": "meGet",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/MeDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND (also outside your scope)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "summary": "Profile, role, scope and permission list (D12).",
        "tags": [
          "me"
        ]
      },
      "patch": {
        "operationId": "meUpdate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/MeDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND (also outside your scope)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "summary": "The caller's own settings; role and scope are refused with 400.",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/devices": {
      "get": {
        "operationId": "meListDevices",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta",
                    "_links"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DeviceDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/OffsetPageMetaDto"
                    },
                    "_links": {
                      "type": "object",
                      "additionalProperties": {
                        "$ref": "#/components/schemas/LinkDto"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "tags": [
          "me"
        ]
      },
      "post": {
        "operationId": "meRegisterDevice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/DeviceDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND (also outside your scope)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "summary": "201 for a new device, 200 when it registers again.",
        "tags": [
          "me"
        ]
      }
    },
...
    "/api/v1/devices": {
      "get": {
        "operationId": "devicesList",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Tables: rows to skip",
            "schema": {
              "minimum": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Feeds: meta.page.nextCursor of the previous page",
            "schema": {
              "maxLength": 512,
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Comma-separated fields, \"-\" for descending",
            "schema": {
              "maxLength": 200,
              "example": "-submittedAt",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search text",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "description": "Comma-separated related resources",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "filter[isDockDevice]",
            "required": false,
            "in": "query",
            "description": "true or false; comma-separated for any of. Also filter[isDockDevice][op] with op in eq, null.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[depotId]",
            "required": false,
            "in": "query",
            "description": "text of up to 200 characters; comma-separated for any of. Also filter[depotId][op] with op in eq, ne, contains, null.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[platform]",
            "required": false,
            "in": "query",
            "description": "one of WEB, PWA, FLUTTER; comma-separated for any of. Also filter[platform][op] with op in eq, ne, null.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta",
                    "_links"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdminDeviceDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/OffsetPageMetaDto"
                    },
                    "_links": {
                      "type": "object",
                      "additionalProperties": {
                        "$ref": "#/components/schemas/LinkDto"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "tags": [
          "devices"
        ]
      }
    },
...
  },
  "info": {
    "title": "Waypoint Compass API",
    "description": "Delivery planning for Waypoint Group (Tech Triathlon 2026). Conventions: specs/api-conventions.md.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Secure-better-auth.session_token"
      }
    },
    "schemas": {
      "UserRole": {
        "type": "string",
        "enum": [
          "admin",
          "dispatcher",
          "loader",
          "driver",
          "store_manager"
        ]
      },
      "MeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "0192a3f4-0000-7000-8000-000000000001"
          },
          "name": {
            "type": "string",
            "example": "Nimesha Periyapperuma"
          },
          "email": {
            "type": "string",
            "example": "nimesha.p@waypoint.lk"
          },
          "username": {
            "type": "string",
            "nullable": true,
            "example": "nimesha.p"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "depotId": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "outletId": {
            "type": "string",
            "nullable": true,
            "example": "OUT014"
          },
          "vehicleId": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "locale": {
            "type": "string",
            "enum": [
              "en",
              "si",
              "ta"
            ],
            "example": "en"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "user:create",
                "user:list",
                "user:set-role",
                "user:ban",
                "user:impersonate",
                "user:impersonate-admins",
                "user:delete",
                "user:set-password",
                "user:set-email",
                "user:get",
                "user:update",
                "session:list",
                "session:revoke",
                "session:delete",
                "order:read",
                "order:create",
                "order:update",
                "order:submit",
                "order:cancel",
                "order:queue",
                "catalog:read",
                "catalog:manage",
                "plan:read",
                "plan:build",
                "plan:publish",
                "plan:revise",
                "plan:close",
                "deferral:read",
                "deferral:decide",
                "deferral:respond",
                "trip:read",
                "trip:reassign",
                "trip:resequence",
                "load:read",
                "load:check",
                "load:flag",
                "load:release",
                "load:decide",
                "stop:read",
                "stop:record",
                "receipt:read",
                "receipt:confirm",
                "issue:read",
                "issue:create",
                "issue:resolve",
                "alert:read",
                "alert:act",
                "tracking:read",
                "forecast:read",
                "forecast:manage",
                "masterData:read",
                "masterData:manage",
                "audit:read",
                "audit:export",
                "settings:read",
                "settings:manage",
                "webhook:manage",
                "simulation:run"
              ]
            },
            "example": [
              "order:read",
              "order:submit"
            ]
          },
          "_links": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/LinkDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "username",
          "phoneNumber",
          "role",
          "depotId",
          "outletId",
          "vehicleId",
          "locale",
          "permissions",
          "_links"
        ]
      },
      "NoticeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "ORDER_ROLLED_TO_NEXT_RUN"
          },
          "message": {
            "type": "string",
            "example": "Sent after the 4 PM cutoff: delivers Saturday."
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "MetaDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "example": "0192a3f5-1b2c-7d3e-8f4a-5b6c7d8e9f0a"
          },
          "serverTime": {
            "type": "string",
            "example": "2026-10-01T15:40:03+05:30"
          },
          "apiVersion": {
            "type": "string",
            "example": "1.0.0"
          },
          "notices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NoticeDto"
            }
          }
        },
        "required": [
          "requestId",
          "serverTime",
          "apiVersion"
        ]
      },
      "LinkDto": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string",
            "example": "/api/v1/orders/0192a3f4-7c1e-7a2b-9d3e-5f6a7b8c9d0e"
          },
          "method": {
            "enum": [
              "GET",
              "POST",
              "PATCH",
              "PUT",
              "DELETE"
            ],
            "type": "string"
          },
          "title": {
            "type": "string",
            "example": "Cancel order"
          },
          "templated": {
            "type": "boolean"
          },
          "requires": {
            "example": [
              "If-Match",
              "reasonNote"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "href"
        ]
      },
      "FieldErrorDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "lines[2].qty"
          },
          "code": {
            "type": "string",
            "example": "min"
          },
          "message": {
            "type": "string",
            "example": "Enter at least 1"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ]
      },
      "ProblemDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "https://compass.waypoint.lk/problems/validation-failed"
          },
          "title": {
            "type": "string",
            "example": "Some fields need attention"
          },
          "status": {
            "type": "number",
            "example": 400
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_FAILED"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/orders"
          },
          "requestId": {
            "type": "string",
            "example": "0192a3f5-1b2c-7d3e-8f4a-5b6c7d8e9f0a"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldErrorDto"
            }
          },
          "violations": {
            "type": "array",
            "description": "PLAN_RULE_VIOLATION only: the broken rules",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "instance",
          "requestId"
        ]
      },
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "locale": {
            "enum": [
              "en",
              "si",
              "ta"
            ],
            "type": "string",
            "example": "si"
          }
        }
      },
      "DevicePlatform": {
        "type": "string",
        "enum": [
          "WEB",
          "PWA",
          "FLUTTER"
        ]
      },
      "DeviceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "P2"
          },
          "platform": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DevicePlatform"
              }
            ]
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "isDockDevice": {
            "type": "boolean",
            "description": "Only dock devices may use PIN sign-in"
          },
          "depotId": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "appVersion": {
            "type": "string",
            "nullable": true,
            "example": "0.1.0"
          },
          "lastSeenAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "platform",
          "label",
          "isDockDevice",
          "depotId",
          "appVersion",
          "lastSeenAt",
          "createdAt"
        ]
      },
      "OffsetPageDto": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "number",
            "example": 10
          },
          "offset": {
            "type": "number",
            "example": 0
          },
          "total": {
            "type": "number",
            "example": 57
          }
        },
        "required": [
          "limit",
          "offset",
          "total"
        ]
      },
      "OffsetPageMetaDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "example": "0192a3f5-1b2c-7d3e-8f4a-5b6c7d8e9f0a"
          },
          "serverTime": {
            "type": "string",
            "example": "2026-10-01T15:40:03+05:30"
          },
          "apiVersion": {
            "type": "string",
            "example": "1.0.0"
          },
          "notices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NoticeDto"
            }
          },
          "page": {
            "$ref": "#/components/schemas/OffsetPageDto"
          }
        },
        "required": [
          "requestId",
          "serverTime",
          "apiVersion",
          "page"
        ]
      },
      "RegisterDeviceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "pattern": "^[\\w-]+$",
            "description": "Generated on the device and kept in IndexedDB",
            "example": "P2"
          },
          "platform": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DevicePlatform"
              }
            ]
          },
          "label": {
            "type": "string",
            "maxLength": 80,
            "example": "Aniqa's phone"
          },
          "appVersion": {
            "type": "string",
            "maxLength": 40,
            "example": "0.1.0"
          }
        },
        "required": [
          "id",
          "platform"
        ]
      },
...
      "AdminDeviceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "P2"
          },
          "platform": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DevicePlatform"
              }
            ]
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "isDockDevice": {
            "type": "boolean",
            "description": "Only dock devices may use PIN sign-in"
          },
          "depotId": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "appVersion": {
            "type": "string",
            "nullable": true,
            "example": "0.1.0"
          },
          "lastSeenAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "_links": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/LinkDto"
            }
          }
        },
        "required": [
          "id",
          "platform",
          "label",
          "isDockDevice",
          "depotId",
          "appVersion",
          "lastSeenAt",
          "createdAt",
          "_links"
        ]
      },
...
    }
  }
}