# packages/api-client/src/gen/model/offsetPageMetaDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Waypoint Compass API
 * Delivery planning for Waypoint Group (Tech Triathlon 2026). Conventions: specs/api-conventions.md.
 * OpenAPI spec version: 1.0.0
 */
import type { NoticeDto } from './noticeDto.ts';
import type { OffsetPageDto } from './offsetPageDto.ts';

export interface OffsetPageMetaDto {
  requestId: string;
  serverTime: string;
  apiVersion: string;
  notices?: NoticeDto[];
  page: OffsetPageDto;
}